Make debugSettings always available in the SceneEnvironment
Create a DebugSettings parented to the SceneEnvironment. If
the user wants to, this can of course be replaced.

Also fix some destroyed signal connections. This was somewhat
wrong for some other properties as well. Make sure we only null
out our reference if the to-be-destroyed object is what we are
associated with, not otherwise.

Change-Id: I02b0bc9f71b4aa405607f1547b56dd2240d64a4e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>