Skip the cubeMultiViewportMultiWindow Test on macOS with OpenGL
Like many other test cases, if you try and use Software OpenGL on macOS
you will not get the expected "real-world" result. At the very least you
would need to force a "core-profile" context which the tests do not do.
OpenGL is also not the default RHI backend on macOS (Metal is), so
issues with using an OpenGL 2.0 context on macOS are expected.  In this
case the test was failing because it could not find a valid shader
compatible with the OpenGL context that it was using, so it wasn't
showing the content in either of the windows.

So because of the limitations of Software OpenGL used for this
particular CI configuration, it can not perform the test so we will only
perform this test on CI configurations that have actual graphics
hardware acceleration.

Fixes: QTBUG-107112
Change-Id: Ife29f9ec1316e2708e31fda0445dbfdc429d709a
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>