Properly fix up the QSGTexture usability check
Now that there is a way to query the creating QRhi from a QRhiResource,
a proper check is possible to implement.

The idea that any QSG resource is usable in any QQuickWindow as
long as the QObject belongs to the same thread as the QQuickWindow
is building on the assumption that only the 'threaded' render loop
uses a different QRhi (and so graphics device/context) for each
window. That was historically true, but never guaranteed.

Task-number: QTBUG-107112
Change-Id: Iaa446e9dc990cd0488acb23da5df0d4150dbc95d
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>