Improve the renderable types
Store the global bounds in the renderable as we'll have use for that
later, e.g., culling. Also, the type was removed from
the renderable flags, as the type will never be reset or changed during
the lifetime of the object. We know the type at creation and is never
going to change. Having a fixed type means it's easier to make sensible
assumption about the object when it's e.g., created.

Task-number: QTBUG-106958
Change-Id: I459734725792b4a7d7fbe4fa88707277e3377b68
Reviewed-by: Christian Strømme <christian.stromme@qt.io>