Add a persistent, disk-based cache for the shader generator
Implemented for default/principled materials and custom materials and
effects. For effects some changes are done to avoid the problem of
colliding hash keys in the first level effect cache. In a somewhat
similar manner, the QQsbCollection is updated to use different kind of
keys (like a string with a hex sha, not a uint64 from qHash).

Relies on a new shadertools API that allows shader baking to skip
target versions that fail in SPIRV-Cross. (because we try to bake
everything, but so have to deal gracefully with things like GLSL ES
100 not having textureSize() etc.)

The cache location is the QStandardPath::CacheLocation, so e.g.
Users\<user>\AppData\Local\<appname>\cache. The .qsbc file is under a
directory q3dshadercache-<buildabi>.

.qsbc files now store the Qt version they were built with.  If there
is a mismatch, the file will be ignored, and upon saving it will be
replaced with new content. (this is identical to how the low-level
pipeline cache works when there is a mismatch)

This is an important change for pregenerated materials: previously any
.qsbc file was loaded, whereas now even a difference in the Qt patch
version will make the prebuilt data to get ignored. This is in theory
not great for apps that build against one version and then replace the
Qt libs with a newer version without rebuilding the app.  Then again
there is no compatibility guarantee for the material shader generator,
so rejecting something from an older version is the right thing to do.

To facilitate this and the key change, the version of .qsbc files is
bumped. We do not support loading the old format.

Eventually an API on some level to control the file location may get
added (just like Qt Quick has QQuickGraphicsConfiguration APIs to
control the low-level pipeline cache). This may be interesting in
particular because the usual caveats apply with multiple windows: the
last window that closes wins, that's the one that gets to write out
its cache. (there is only a single file) The story here is identical
to Qt Quick: If one wants to have proper control over this, they need
to explicitly specify the file location on a per-window basis (not per
View3D, though, which may complicate things).

The persistent caching can be disabled using the existing mechanisms,
meaning disabling the Qt Quick pipeline cache (or the Qt 5 era
QOpenGLShaderProgram disk cache) will also disable this one. One can
set Qt::AA_DisableShaderDiskCache, or the env.var
QT_DISABLE_SHADER_DISK_CACHE or its QSG_RHI_... equivalent. This way
we do not need a Quick3D-specific setting at all.

For development purposes it can be handy to disable just the loading
of a possibly exising cache, while keeping the saving intact
(i.e. this allows overwriting any existing cache data when changing
the material shader generator, without running into problems due to
loading cached data from a Quick3D build without the latest
changes). To do this, set QT_QUICK3D_NO_SHADER_CACHE_LOAD to a
non-zero value.

To confirm what is happening under the hood, set the usual
QT_RHI_SHADER_DEBUG env.var. This will cause:

- upon QSSGShaderCache construction the printing of something like the
  following, if there is an existing cache:

Attempting to seed material shader cache from <filename>
Loaded 15 shader pipelines into the material shader cache

- then on each default/custom material lookup it shows one of:

   * nothing (if the highest level, runtime cache were hit, i.e. we
     saw the material before during the run of the application),

   * or it prints something along the lines of "Loading rhi shaders
     from disk cache for <cache key> (<material key>)". That is when a
     new QSSGRhiShaderPipeline was created on the cheap path,
     i.e. when the QShaders are coming from the .qsbc file.

   * Finally, if it prints the shader source code then it's on the
     slow path with generating and processing the shaders.

A dedicated documentation page will likely need to be introduced
separately. That is however postponed until we get an idea if there
will be dedicated APIs to control the cache location or not.

Change-Id: I166f72c3f2f4ea0fbcf0b0af658117cdc537b241
Reviewed-by: Andy Nichols <andy.nichols@qt.io>