Prevent unnecessary memory allocation per frame
The texture size by of QSGPlainTexture is (-1, -1) by default and it
is not set when calling setTexture. It is used to compare current
render size to previous frame texture size, which always fails and
new QSGPLainTexture gets allocated every time. Set the texture size
manually so that it gets compared to the correct size.

Pick-to: 6.4
Change-Id: I661a83d90e30489e86cdb05e404e14778d944209
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>