Also release shadow and reflection map resources when requested
These are recreated automatically when rendering the next frame.
It is however very useful that we can empty these lists since if the
shadow/reflection state is toggled dynamically, the allocated textures
stay around for potential future reuse. This is not always great so
now calling window->releaseResources() or clicking the button in the
DebugView also purges these. The effect is very visible in the 'lights'
example for instance: disabling shadows with the checkbox will not
lead to a drop in the graphics memory allocated, but clicking on the
Release Resources button will now.

Also unify the naming: we always call it release*Cached*Resources() now.
Otherwise there is an inevitable confusion with QQuickItem's
releaseResources() which is not related to this.

Change-Id: I937fc34ad7d33e27f9fe7b8d959b2b8b929a3630
Reviewed-by: Andy Nichols <andy.nichols@qt.io>