CMake: Fix "command too long" issue when building using MinGW windres
CMake doesn't support creation of response files for .rc resource
compilers.
https://gitlab.kitware.com/cmake/cmake/-/issues/16171

The asset importer plugin has many include directories which are
passed to windres.exe directly on the command line and that ends up
exceeding the command line character limit.

The simplest work around is to disable the compilation of the .rc file
for the plugin when building using MinGW.

Other workarounds were considered but did not work.

Restricting the include directories using the $<COMPILE_LANGUAGE>
generator expression doesn't work because it causes other issues for
documentation generation.

Previously we used object libraries so that include directories are
not inherited, but that caused issues with newer CMake versions, as
can be seen in the implementation of
_qt_internal_generate_win32_rc_file.

Pick-to: 6.4
Fixes: QTBUG-96339
Fixes: QTBUG-98336
Fixes: QTBUG-107734
Change-Id: I79b99396e6241ac3bc3efb2f24d1762e2d24ab1c
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>