Improve Effect shader key calculation
...to capture the variation we perform for the vertex shader code.

If the source code is varied based on rhi->isYUpInFramebuffer(),
then this must be captured in the cache key as well.

Otherwise we end up in an unfortunate situation where running the
same application with a different rhi backend (and 3D API) gives
incorrect results if the 3D APIs in question differ in the return
value from isYUpInFramebuffer, since we incorrectly pick an existing
entry from the disk cache.

This was not relevant until persistent caching was introduced, since
the rhi backend (3D API) could only be changed by restarting the
application.

Change-Id: I1c1e0ad80809dc2612575c4bd88b1263c96030b1
Reviewed-by: Christian Strømme <christian.stromme@qt.io>