Small cleanup of fragment shader light generation code
During my attempt to add independent light count support to  the light
calculations in QSSGMaterialShaderGenerator, I managed to cleanup/
refactor parts of the light calculation a little bit. Because of the
uncertainty of that other patch, this is a patch with just that cleaned
up code, hopefully making the code a little bit easier to follow and
work with.

Change-Id: If305a7330e1385556f9fef65c69f5608d70a4d54
Reviewed-by: Andy Nichols <andy.nichols@qt.io>