Fix file path resolution for BakedLightmap
The loadPrefix and file names of the mesh and exr files was never
resolved, so local paths would be interpreted as relative to the app's
current working directory, instead of the common behavior: relative to
the qml source file.

This fixes the rendering of the lighting/lightmap.qml lancelot scene.

Change-Id: I1d9af487133698550233c560c50da1a20e3256ee
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 0f51037784881227f7b8b554dbe080b6f8e1bce2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>