Improve RenderStats frame time accuracy
Using endFrame() as the universal starting and ending point for
measuring frameTime is not very accurate since many times we do not
render every frame on purpose. To get a good idea of how long a frame
really is, we need to restart the frame timer every time we present a
frame because we wont know until later if we are going to render for
that frame. To account for this, we need access to the QQuickWindow
which we are presenting through so that we can get a signal whenever a
frame is presented. That way we know the true "End" of a frame,
because GPU commands were executed and presented, and we know the
"Start" of what may be the next frame so we have something to measure
against.

One controversial thing about this change is the FPS measurement.
What is measured in the RenderStats is FPS while rendering, not true
frames placed on screen per second. This is meant to account for times
when content is not being updated. If nothing in the scene changes,
then with true FPS you would get an FPS of 0. If you update a few
times because you moved your mouse for half a second, you would get
say 30FPS, even though your rendering was completely smooth with not
frames dropped (you just only needed to render 30 times). If you want
"true" FPS then you should measure the whole window. I think measuring
only rendered frames and their timings in RenderStats is more
meaningful because half a second of movement with no frame drops would
display the intended measurement (60 FPS for a 60Hz screen). So this
is still a real FPS measurement, it is just not the absolute per
second of the whole window.

Change-Id: I6429ea2c1d56b75876c758ed31892bd8ea416334
Reviewed-by: Christian Strømme <christian.stromme@qt.io>