Fix up the tables in DebugView
Use TableViews with real models, not just a Text field.

Compromise and use a non-resizable header for the tables for now.
Attempt to use column stretch factors based on the overall width of
the debug panel, so the tables fill the width, while the scrollbar
is still placed and sized sensibly (may not be perfect with native
styles but should do for now). The end result looks pretty good
in fact, but there may be some gotchas depending on font sizes
or styles.

Further optimizations may come separately. Parsing the markdown
strings generated by renderstats is not ideal, for example. (but
working with a single string containing all the data is fairly
handy in practice) To be checked separately if eliminating that,
and rather directly updating the models somehow, or using a more
lightweight text format instead of the markdown tables (which
are only used due to historical reasons anyway) is sensible.

Fixes: QTBUG-109814
Change-Id: I8dcece92550da43b36c6f7cf3e850a73f1ce86e5
Reviewed-by: Kristoffer Skau <kristoffer.skau@qt.io>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 0cddb97fa21be67df468ca9be78cac7e275acf3c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>