Update occlusionMap docs to reflect reality
Due to the patch mentioned below the occlusion map is
effective for any sort of lighting, both direct and
indirect (IBL). This is also why the relevant Lancelot
test gives different results now.

If this approach is the right thing to do, then the
documentation should be changed not to talk about indirect
lighting, but lighting in general.

Amends 5e4e7e2547b863525c6ff51bbc432fc9db648af1

Change-Id: I64faa788de3dbfb9f016029878138afcf5870a5a
Reviewed-by: Christian Strømme <christian.stromme@qt.io>