Clean-up some asserts
Remove anti-pattern of adding asserts in the default cases' of switch
statements that are "complete". The compiler can and should fail if not
all types are specified. Relaying on missed cases being "found" at
runtime, eventually, is not ideal and we're also not getting any help
from the compiler/tools as to which switch statements needs to be
modified when new types are added. The goal of this change is therefore
to utilize the compiler to help with e.g., compile time checks and also
for us to help the compiler by telling it about code paths that should
be unreachable.

Change-Id: Ia72d237c7aa4e79e6358a82742f601f481b2796e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 5c53152e4022c1dac242e67d500a027279958758)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>