Trigger lightmap baking interactively from DebugView
Currently the only way to bake lightmaps is to launch the application
with --bake-lightmaps or use
a specific environment variable. When baking ends, the application will
exit. This is not really ideal for development.
A solution to aid with this is a new button in DebugView so that we have
an interactive way of triggering the baking. When baking is complete the
lightmapper will try to save to the same location as the loaded assets
to overwrite them. This plus a refresh of resources when done provides
live reloading. This only works if the files are not bundled with the
executable. If they are, the newly baked files will be put in the
working
directory.
This same mechanism might be used in Design Studio down the line to
provide lightmap baking from there as well.

There's currently no feedback being provided in DebugView, this will be
handled in a later patch. For now the console debug logs will do.

Fixes: QTBUG-109815
Change-Id: I7a050638a2a3fcb00d7b63c518b4756b1ca0f0f2
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>