Interactive lightmap baking improvements
Expanded the callback that's provided to QSSGLightmapper for runtime
lightmap baking. It now returns an enum with type and an optional string
message. A third parameter is a pointer to a struct that's provided for
communication with the mapper during baking. For now it just contains a
flag for requesting a cancel operation. This flag is checked at certain
points during baking, and will abort if set.

The patch also contains a slight logic change to the qlm_list file. It
will now always be written to the working directory, but the file paths
it contains will be absolute. This is to account for the fact that
loadPrefixes can in practice be completely different and make it hard to
for the denoiser to work with.

Bakedlightmap example is slightly reworked to show that the preferred
way to bake lights is to use the DebugView button. All exr files are
removed from the bundle so that baking is required to see the correct
result. This makes the example pretty useless on platforms that does not
support lightmap baking, but it does make sense why.

Pick-to: 6.5
Change-Id: I72f48780e0c3983d6a1941db1b833b78784b99dd
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>