Make WindowAttachment to be destroyed after window
If WindowAttachment is a child of window, it is destroyed
during destroying window.
However, it should be deleted after all the cleanups related
with 'window', otherwise some resourses deleted after it,
will not be cleaned correctly.
For example, when rendering a multi view, the first view makes
a WindowAttachment and it becomes the next child after the view.
However, window can make other views and they might be children
after the WindowAttachment. When cleanupNodes, all the views
except the first view might be destroyed after the
WindowAttachment was already deleted. And this will lead to
fail the cleanup process.

Change-Id: I316bafd46678b3b2b52ada2d855dc9f594000226
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 322540c9350977da16de59a9e734a7005c33cd55)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>