API Fixes: ExtendedSceneEnvironment texture properties
Previously none of the texture properties were done correctly:
For LUT color grading, we gave access the the url of a built in texture
which was inconsistent with the rest of the texture APIs
For the Lens Flare Dirt, Starburst, and Color textures the property was
pointing to the TextureInput property and not to the Texture inside, so
it was not plumbed as intended.  This has been fixed, and now all
texture properties accept a Texture component.

Change-Id: I653df27b1ff0860883dd12d32ec08466bac9940e
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 341c8fa9f793325c100407afc463309e6289e3a2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>