Make lights example ui to more mobile friendly
Changed settings to drawer that can be opened and closed, so that it no
longer can block the view. Added scrolling to settings, so that all the
settings can be used when running the example on smaller screens.

Fixes: QTBUG-109992
Change-Id: I404a88ded7073483d1a96de4e108fa60aba113f8
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 1ed813030c87c62dd7921a146ca56541735eb0b6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>