Fix problem with the bounding volume calculation for particles
The bounding volume was wrongly calculated which caused problems when
doing frustum culling and shadow map inclusion. The issue is that we
need to have special handling for model blend particles as the bounding
volume is already in global space, and this handling was broken when
the renderable type handling was changed in 4130d3d31a356.

Fixes: QTBUG-110580
Task-number: QTBUG-96184
Change-Id: I94c1dcb138501aa771d4dbcd2ce1de81940b4677
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit a916058c6eea34aa641a68ebd1d31211776ab02e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>