Fix CubeMapTexture relative path handling
Using a semicolon-separated list of relative paths would break when
reading from the file system if the current directory was different from
QQmlEngine::baseUrl(). This is a quick fix: resolve each path in the
list separately. In the long term, there should be support for a list of
paths in QSSGRenderImage.

Fixes: QTBUG-110962
Pick-to: 6.5
Change-Id: I3473ad78d892c678cea6ff4411b401bbab91dffc
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Inho Lee <inho.lee@qt.io>