Mark types in QtQuick3D.Effects as deprecated
Note that the Effect type itself, that is used to implement custom
postprocessing effects, is not part of this module and is not
deprecated. Only the premade effects inherited from NDD/Qt 3D Studio
are made obsolete here. The documentation now suggests using
ExtendedSceneEnvironment, MultiEffect, or the application's own custom
effects instead.

[ChangeLog][QtQuick3D.Effects] The pre-made post-processing effects
inherited from Qt 3D Studio and its predecessors are now marked as
deprecated. They are still available for use, but new projects are
encouraged to use the 3D post-processing facilities provided by
ExtendedSceneEnvironment and SceneEnvironment. Whereas 2D effects
that operate on and treat the output of the View3D as a simple 2D
image, are recommended to be implemented via Qt Quick's facilities,
such as the new MultiEffect type or the Qt Quick Effect Maker tool.

Pick-to: 6.5
Change-Id: I90c836023d62da5250dc8142d3db2f840e1e41ad
Reviewed-by: Andy Nichols <andy.nichols@qt.io>