Lights, don't copy the light data to each node entry
Each node entry will contain a reference to the global light list
unless there's a scoped light attached to the node, in which case
it will get a bespoke light list. In addition all non-lighting
lights are culled in the prep phase, that means after the prep
phase there's no lights with brightness <= 0.

Pick-to: 6.5
Change-Id: Ida14abbfa306a735636b120339204b3634159c73
Reviewed-by: Christian Strømme <christian.stromme@qt.io>