Update handling of QUrl Texture source property in valueToQml
When looking for textures and the property value is a QUrl there are no
output showing when errors occur and the texture source property value
in the final .qml is wrong.
The source property value in the .qml ends up being the path where it
tried to find the original texture prior to copying to the final output
directory, which is misleading. With this patch it's updated to instead
set ok to false (which results in adding the line Skipped property:
source to the qml file) and create a comment with the final expected
path (e.g. maps/x.png). This allows the user to create the maps folder
and add the textures themselves if wanted. If they want to fix the
problem beforehand, so before running Balsam, there are now also output
logs when running Balsam indicating that it failed to find the textures
with the absolute path of where it expected to find them prior to
copying into the final output directory. This allows the user to move
the textures to the correct place for the import tool to find them.
There's also a small bug, when texturePath is a complete valid absolute
path, calling path() results in a no longer valid path, e.g. on Windows
it removes anything prior to the colon, like the drive letter. This is
now updated to just use the string as is if it's absolute.

Task-number: QTBUG-110918
Pick-to: 6.5
Change-Id: If4cc1bc026773523a8caf0331b3e186f1ba5bba8
Reviewed-by: Christian Strømme <christian.stromme@qt.io>