Make sure models with an instanceRoot is updated properly
Prioritize models with an instance root so they are processed later
than the instance root to avoid that the transforms gets messed-up.
Ideally we should have had a dedicated type for the instance root,
but for now (or instead) we workaround that by modifying the priority
manually, which also means we have an actual use-case for the dirty
type, as we need to re-evaluate the node priority when it's instance
root changes.

Pick-to: 6.5
Fixes: QTBUG-110996
Change-Id: I650670d64170c4fae3d6aeb8d81317431bb171e4
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>