Skip unneeded calls to rhiPrepareSkyBox
Do it only once per frame (per layer, as in View3D).

Due to the existing function mixing together two different things
(plain skybox, skybox for reflection map), and already being called
multiple times potentially, the best is to separate the two cases
and use the flag to skip subsequent calls for the "plain" version
only. The other one (reflection map/probe) is called multiple times
by design already.

We check and update the layerrenderdata flag in the caller, not in the
rhiPrepare* functions themselves. This is because no other similar
function has such skip-if-something logic, so for clarity this is left
to the callers.

Task-number: QTBUG-111357
Change-Id: Ic2fd8152012bd0dc6cc2211a3cdecb11f07d6f69
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit e1eccadba33e938fd34453c128fa8778b7b98451)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>