Update internal input processing of QQuick3DViewport
Right now internal input handling is always enabled. This causes
unnecessary performance issues, as this should only be enabled if there
are any 2D stuff that require it, so either a Item2D or a Image2D with a
sourceItem.

This is accomplished by adding a counter to the scene manager that holds
the currently active number of Item2Ds in the scene. If this number is
greater than zero, we implicitly enable internal input processing in
QQuick3DViewport.

Also added an environment flag that overrides this, and makes the
behavior the same as it used to, which sets the flag permanently to true
in the constructor, and won't toggle it on and off at runtime. This is
mostly to help with the input tests, as this new behavior seems to
happen in the wrong order for the tests to pass.

Fixes: QTBUG-111276
Change-Id: I510d28e2dbcdf2c2e5ce85e08fc05778b36358b8
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 794e179c0f427b43aad3b9728b65d5131469f77d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>