Asset: fix to handle embedded textures
Embedded textures are not always used by "*" + number format.
Since GetEmbeddedTextureAndIndex in assimp supports to get index
and name at the same time, it will be used instead of previous
logic.

If an embedded texture has a compressed format, the format hint
should be delivered to QImageReader.
TextureData has a member, 'fmt', related to the image format,
but it was not used. It will be used as the format hint, so
the type is changed to QByteArray.

Fixes: QTBUG-110918
Pick-to: 6.5
Change-Id: I33cc7871073bca383fc007dd8eaac38ebe7910ab
Reviewed-by: Christian Strømme <christian.stromme@qt.io>