Add more QQuick3dViewport picking alternatives with mouse coords
Currently when picking with viewport coordinates a user can only choose
between a version that returns the closest one, and one which returns a
list of all intersected objects. What these both have in common is that
they iterate through the entire scene, and checking every single one
that is pickable or actually all if global picking is enabled. This
patch adds a couple of variations that is an optimization to this.

QQuick3DViewport::pick has a new overload where you can pass in a
QQuick3DModel and it will only check for intersectios with that specific
one. Added picking auto test additions for this.

There's also a new function called QQuick3DViewport::pickSubset, which
allows you to pass in a list of QQuick3DModels and get a list of hit
results based on intersections only with those. No picking auto test
additions here, because this function is only really usable from QML.

There's a new manual test that shows all this in practice. Also serves
as a small example of how the subset can be dynamically updated at
runtime if required.

[ChangeLog][Quick3D][QQuick3DViewport] Added additional picking
alternatives that allows for dynamically checking against a subset of
the models in a scene.

Fixes: QTBUG-111700
Change-Id: I51964a2cb9f3d7badb5fdc47b406754ac26886fe
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>