Calculate Triangle Intersection using Möller–Trumbore algorithm
This change refines the previous approach for handling triangle ray
intersection to be more effient and cover more corner cases. The
algorithm being used now should be pretty close to what is described in
the referenced paper and follow up optimization of the paper's example.

This should fix issues that we saw previously with very small triangles
due to floating point error.

Additional tests were added to verify the small triangle case as well as
making sure that intersections occurred on triangle corners. There was 2
tests that needed some adjustment since some assumptions about the test
data was wrong (specifically moving the origin point of the ray out of
the plane of the triangle under test).

Fixes: QTBUG-111997
Pick-to: 6.5
Change-Id: Iede12c7d2d51787e8c3ad3e35548e63d867e7142
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>