Improve how we trigger updates for multi-view multi-window cases
Only do an inline sync when the imported scene has an RCI and
the underlying RHI is the same. If there's no RCI for the
imported scene, request an update so one gets set.
Note that with this change the auto test was changed to not
expect a warning from the buffer manager, as we shouldn't
call with a incompatible RHI

Pick-to: 6.5
Change-Id: Id5ee703b6a079b851974faef9f02e2b2ade7e1b2
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>