RenderContextInterface improvements
Move the RCI to the window attachment and get rid of the old global
storage. The RCI is per-window, same as the window attachment, so
it make sense to keep them together. The releaseCachedResources is
also tied to the window attachment, which means we can have a more
flexible callback system, for when cached resources should be cleared.

Pick-to: 6.5
Change-Id: If32318cdc492ed2303062523a969395706c84a4e
Reviewed-by: Christian Strømme <christian.stromme@qt.io>