Fix handling of empty scene shadow-casting bounds
Commit 65bc16ad84c3d94446ca4b0da8b798593d1241ae introduced a change
where the conversion to QSSGBoxPoints set the objects' bounding box to
zero instead of "empty" in calculateSortedObjectBounds.

This change will revert so we get an empty bounding box instead of a
zeroed one since that is what the calling code expects. We also
refactor the QSSGBoxPoints conversion method to be more readable.

Also adding a lancelot test and a manual test.

Pick-to: 6.5
Fixes: QTBUG-112450
Change-Id: I32831680c1f356ea58b1b0be5861e5c30bdc99c6
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>