inline sync change
Be a bit less strict for when we trigger inline syncs, if the views'
RHI instances are on the same thread we're still ok to proceed, e.g.,
when using the basic renderloop. For windows that are not exposed
we also assume we can do a sync of the views nodes, which we
haphazardly did previous versions anyways (note that the view's window
will need to be exposes at least once to have it's context initialized).

Pick-to: 6.5
Change-Id: Ifc7ee9102197fb608d8171aaf56b74c1f769b134
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>