Extract the prep steps for meshes and bone textures
Doing the mesh loading prior to the prep allows us to more easily filter
out bad models (with missing meshes) early. The prep step for setting up
the bone texture is an optional step which we should aim at avoiding if
possible. The three functions still maintain a strong dependencies and
needs to be called in sequence.

Change-Id: I0d36e7e746efbf4c837e6aa503245160dd925a43
Reviewed-by: Andy Nichols <andy.nichols@qt.io>