Remove dynamictexture example
While this example had some use as a way to demonstrate input
interaction with 2D items in a 3D scene, it did not do so in a very
effective way. The code for the 2D part came from an old QtQuick demo,
and did not demonstrate very modern QML usage. Instead the demo and code
looked very dated, and does not do a very good job of simply showing how
to create a "dynamic" texture from Items based components.

So this commit removes the example and demotes it to a manual test.

Change-Id: I65346c7b621c65e1096e1c6cd2638a14853efe17
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit cea29f96fa5fdf932916457272bbe1bb262dafa9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>