Add function to retrieve render results from other passes
This means there's no need to know which pass produced the result and
pull it out from there, it's instead available through the frame data
class. For now this is only for the built-in results, but we can and
should have a similar mechanism for extensions as well.

Change-Id: Id6bfa88a4582c871bcf34405299a378af4010d85
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 028840c53d7d7ab14f53fa9f2acfee2e3eb53d15)
Reviewed-by: Christian Strømme <christian.stromme@qt.io>