Have a way to disable auto save/load in QSSGShaderCache
This needs to be static since a client, such as the QML Puppet,
has to be able to state that automatic load/save is not wanted
before the loading happens, and that happens in the QSSGShaderCache
constructor.

The automatic disk loading can be accessed through the
QtQuick3DEditorHelpers namespace.

Change-Id: I5481239e610657119081a94f187a34897d463d25
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 7148128f1f331c2306ad549f97b133a280fe9299)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>