Make sure the "bounds" type is registered with QML
The QQuick3DBounds3 type has been around since 5.15 and is used in two
API's (Model, and RuntimeLoader), but did not actually have a QML type
defined for it. Documentation was in place for the "Bounds" type but that
type never existed, and was the wrong name since QQuick3DBounds3 is a
value type.  The documentation has been corrected, and the "bounds" type
is now registered such that the type will show up in qmltypes.

This is not an API change, everything works as before, just you should
now have the type defined for type checking in QML for QQuick3DBounds3.

Change-Id: I46273eddcfde886c35b981e5394467ebe9e2bc65
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 3bd1aa0adcacc707f3b6799dc212b5fbc098e9dc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>