Balsam: Add manualAnimations option on import
This makes it where TimelineAnimations are not generated along with the
Timeline components. The automatic generation of TimelineAnimations was
so that we would always play some animation if one was available, but
for assets with many animations it is likely that a more manual approach
is required. So when using --manualAnimations just the Timeline(s) are
created as well as a mapping between the original asset name (if
available) and the created Timeline.

I also refined the behavior of how the Timeline objects are named. For
the "id" values, this must always be sanitized since it has to be both
unique and a valid QML id. The "objectName" property however is more
flexible, and if the asset provided a string name for this, that is what
the object name will be.  If there is no value, then the objectName will
just be the sanitized id. The idea is that that there should be some way
to map the original name given in the DCC tool to the QtQuick based
instance.

Change-Id: I8429340b6f65c467e4686ad76886090e4e3ba513
Reviewed-by: Inho Lee <inho.lee@qt.io>
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>