Support HDR (linear) images as base color maps
This can be seen in action in the hdrq3d manual test for now, or
everywhere where a .hdr or .exr file is used as the base color map.

[ChangeLog] Using HDR images (loaded from .hdr or .exr files) as base
color maps are now supported correctly. Previously the image data was
incorrectly going through sRGB->linear conversion which is incorrect
for HDR image data as they are not sRGB (or any other color space
for that matter).

Change-Id: I693e4128252b408ecc4717d857b82ff532b18eb1
Reviewed-by: Andy Nichols <andy.nichols@qt.io>