Mark the module free of Q_FOREACH ... again
This restores the QT_NO_FOREACH macro originally added in
03dbcc5eb3364886ee8ffc07b4438eaee3c188fe and then lost in the cmake
port.

Fix the new uses that have crept in in the meantime. They're all
trivially safe as ranged for loops.

As a drive-by, re-use the const temporary added by this patch in
subsequent statements instead of re-fetching from the source. Take
advantage of its const-ness and fix the coding style issue (operator!
spacing).

Change-Id: I0416f41e23e22e3f80d32ba68d46612bafccf234
Reviewed-by: Christian Strømme <christian.stromme@qt.io>